# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentGuard Lite API",
    "version": "1.0.0",
    "description": "Frozen HTTP contract for signalling, AgentGuard execution, ledger inspection, and the React client. No server URL is embedded; the frontend receives it from runtime configuration."
  },
...
  "paths": {
...
    "/api/rooms": {
      "post": {
        "tags": ["Rooms"],
        "operationId": "createRoom",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "./schemas/room-manifest.schema.json"}
            }
          }
        },
        "responses": {
          "201": {
            "description": "Immutable room manifest stored",
            "content": {
              "application/json": {
                "schema": {"$ref": "./schemas/room-manifest.schema.json"}
              }
            }
          },
          "409": {"$ref": "#/components/responses/Conflict"},
          "422": {"$ref": "#/components/responses/ValidationError"}
        }
      }
    },
    "/api/rooms/{room_id}": {
      "parameters": [
        {"$ref": "#/components/parameters/RoomId"}
      ],
      "get": {
        "tags": ["Rooms"],
        "operationId": "getRoom",
        "responses": {
          "200": {
            "description": "Exact stored room manifest",
            "content": {
              "application/json": {
                "schema": {"$ref": "./schemas/room-manifest.schema.json"}
              }
            }
          },
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
    "/api/rooms/{room_id}/members": {
      "parameters": [
        {"$ref": "#/components/parameters/RoomId"}
      ],
      "post": {
        "tags": ["Rooms"],
        "operationId": "joinRoom",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "./schemas/room-join.schema.json"}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Membership accepted or refreshed",
            "content": {
              "application/json": {
                "schema": {"$ref": "./schemas/room-member.schema.json"}
              }
            }
          },
          "404": {"$ref": "#/components/responses/NotFound"},
          "422": {"$ref": "#/components/responses/ValidationError"}
        }
      },
      "get": {
        "tags": ["Rooms"],
        "operationId": "listRoomMembers",
        "responses": {
          "200": {
            "description": "Non-expired room members",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["items", "server_time_ms"],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {"$ref": "./schemas/room-member.schema.json"}
                    },
                    "server_time_ms": {
                      "$ref": "./schemas/common.schema.json#/$defs/timestampMs"
                    }
                  }
                }
              }
            }
          },
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
    "/api/rooms/{room_id}/members/{node_id}/heartbeat": {
      "parameters": [
        {"$ref": "#/components/parameters/RoomId"},
        {"$ref": "#/components/parameters/NodeId"}
      ],
      "post": {
        "tags": ["Rooms"],
        "operationId": "heartbeatRoomMember",
        "responses": {
          "204": {"description": "Heartbeat accepted"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
    "/api/rooms/{room_id}/members/{node_id}": {
      "parameters": [
        {"$ref": "#/components/parameters/RoomId"},
        {"$ref": "#/components/parameters/NodeId"}
      ],
      "delete": {
        "tags": ["Rooms"],
        "operationId": "leaveRoom",
        "responses": {
          "204": {"description": "Membership removed"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "RoomId": {
        "name": "room_id",
        "in": "path",
        "required": true,
        "schema": {"$ref": "./schemas/common.schema.json#/$defs/roomId"}
      },
      "NodeId": {
        "name": "node_id",
        "in": "path",
        "required": true,
        "schema": {"$ref": "./schemas/common.schema.json#/$defs/nodeId"}
      },
...
    },
    "responses": {
      "NotFound": {
        "description": "Requested resource does not exist",
        "content": {
          "application/json": {
            "schema": {"$ref": "./schemas/common.schema.json#/$defs/apiError"}
          }
        }
      },
      "Conflict": {
        "description": "Request conflicts with current state",
        "content": {
          "application/json": {
            "schema": {"$ref": "./schemas/common.schema.json#/$defs/apiError"}
          }
        }
      },
      "ValidationError": {
        "description": "Schema or domain validation failed",
        "content": {
          "application/json": {
            "schema": {"$ref": "./schemas/common.schema.json#/$defs/apiError"}
          }
        }
      },
...
    },
...
  }
}